Extract Telemetry methods
